Add google sign up form instead of slack link
This allows us to batch up invites, and have some more control over
the onboarding process, there's a greater chance of people being around
to welcome new members, or chat to others.

Admins an still add users manually to slack, if we don't want to wait
til the next Monday.